first project of 42 cursus done and tested with bonus

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: faksouss <faksouss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 18:58:31 by faksouss          #+#    #+#             */
-/*   Updated: 2022/10/11 01:29:08 by faksouss         ###   ########.fr       */
+/*   Updated: 2022/10/13 23:40:59 by faksouss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
